feat: confirm complete

+++ src/views/layout/components/header/header-search/history.vue
@@ -29,6 +29,7 @@
 </template>
 <script setup>
 import { useStore } from 'vuex'
+import { confirm } from '@/libs'
 
 const EMITS_ITEM_CLICK = 'itemClick'
 
@@ -40,7 +41,9 @@ const store = useStore()
  * 删除所有记录
  */
 const onDeleteAllClick = () => {
-  store.commit('search/deleteAllHistory')
+  confirm('提示', '要删除所有历史记录吗？').then(() => {
+    store.commit('search/deleteAllHistory')
+  })
 }
 
 /**
